last update

+++ src/routes/categories.js
@@ -5,16 +5,24 @@ const { body, validationResult } = require('express-validator');
 // METHOD GET
 
 // GET ALL CATEGORIES
-categoriesRouter.get('/all', async (req, res, next) => {
+categoriesRouter.get('/count', async (req, res, next) => {
 	try {
 		let categories = await db.Categories.findAll();
+		res.status(200).json(categories.length);
+	} catch (error) {
+		res.status(500).json(error.message);
+	}
+});
+categoriesRouter.get('/', async (req, res, next) => {
+	try {
+		let categories = await db.Categories.findAll({ order: [ [ 'categoryID', 'DESC' ] ] });
 		for (let i = 0; i < categories.length; i++) {
 			let attributesCategory = await db.Attribute.findAll({
 				attributes: { exclude: [ 'categoryID' ] },
 				where: { categoryID: categories[i].categoryID }
 			});
 			categories[i].dataValues.attributes = attributesCategory;
-			console.log(categories[i].dataValues);
+			categories[i].dataValues.id = categories[i].dataValues.categoryID;
 		}
 		res.status(200).json(categories);
 	} catch (error) {
@@ -33,6 +41,8 @@ categoriesRouter.get('/:categoryID', async (req, res, next) => {
 				where: { categoryID: category.categoryID }
 			});
 			category.dataValues.attributes = attributesCategory;
+			category.dataValues.id = category.dataValues.categoryID;
+
 			res.status(200).json(category);
 		} else res.status(400).json('category not exist!');
 	} catch (error) {
@@ -54,10 +64,15 @@ categoriesRouter.post('/', body('name').notEmpty(), body('attributes').isArray({
 	try {
 		let category = await db.Categories.create({ name });
 		attributes.forEach(async (attr) => {
-			await db.Attribute.create({ categoryID: category.categoryID, name: attr });
+			await db.Attribute.create({ categoryID: category.categoryID, name: attr.name });
 		});
-		res.status(201).json({ ...category.dataValues, attributes });
+		let attributeCategory = await db.Attribute.findAll({ where: { categoryID: category.categoryID } });
+		category.attributes = attributeCategory;
+		category.dataValues.id = category.dataValues.categoryID;
+		console.log(category);
+		res.status(201).json(category);
 	} catch (error) {
+		console.log(error);
 		return res.status(400).json('category name exited!');
 	}
 });
@@ -65,27 +80,25 @@ categoriesRouter.post('/', body('name').notEmpty(), body('attributes').isArray({
 // UPDATE NAME OF CATEGORY
 categoriesRouter.put('/:categoryID', async (req, res, next) => {
 	let categoryID = req.params.categoryID,
-		name = req.body.name;
-
-	let category = await db.Categories.findByPk(categoryID);
-	if (category) {
-		if (name) {
+		name = req.body.name,
+		attributes = req.body.attributes;
+	try {
+		let category = await db.Categories.findByPk(categoryID);
+		if (category) {
 			await db.Categories.update({ name }, { where: { categoryID } });
-		}
-		// if (attribute) {
-		// 	let attribute = await db.findOne({ where: { categoryID, name: attributes[i] } });
-		// 	if (attribute)
-		// 		await db.Attribute.update(
-		// 			{ name: attribute[i] },
-		// 			{ where: { attributeID: attribute.dataValues.attributeID } }
-		// 		);
-		// }
-		let attributesOfCategory = await db.Attribute.findAll({
-			attributes: { exclude: [ 'categoryID' ] },
-			where: { categoryID: 8 }
-		});
-		res.status(201).json({ categoryID, name, attributes: attributesOfCategory });
-	} else return res.status(400).json('Category not exist!');
+			await db.Attribute.destroy({
+				where: { categoryID }
+			});
+			for (let attr of attributes) {
+				await db.Attribute.create({ name: attr.name, categoryID });
+			}
+			category.dataValues.id = category.dataValues.categoryID;
+			res.status(201).json(category);
+		} else return res.status(400).json('Category not exist!');
+	} catch (error) {
+		console.log(error);
+		return res.status(400).json(error);
+	}
 });
 // METHOD DELETE
 categoriesRouter.delete('/:categoryID', async (req, res, next) => {
@@ -93,10 +106,13 @@ categoriesRouter.delete('/:categoryID', async (req, res, next) => {
 	let productOfCategory = await db.Product.findOne({ where: { categoryID } });
 	if (productOfCategory) return res.status(400).json('exited product using category!');
 	try {
+		let products = await db.Product.findAll({ where: { categoryID } });
+		if (products.length > 0) return res.status(400).json({ ...req.body });
 		await db.Attribute.destroy({ where: { categoryID } });
 		await db.Categories.destroy({ where: { categoryID } });
-		res.status(204).json('removed category!');
+		res.status(200).json({ ...req.body });
 	} catch (error) {
+		console.log(error);
 		return res.status(400).json(error.message);
 	}
 });
